worker/requirements.txt
google-cloud-pubsub
openai
pydub
python-dotenv
supabase
sentence-transformers
pdfminer.six
python-docx
openpyxl
pandas
fastapi
uvicorn
google-cloud-aiplatform[generative_models]
